removed middleware, changed the rules for deleting a comment and added the endpoit to hide the comments.

+++ src/controllers/comment/deleteComment.controller.ts
@@ -1,7 +1,8 @@
-import { Request, Response } from "express";
 import Joi from "joi";
+import { Request, Response } from "express";
 import { Comment } from "../../schemas/comment.schema";
 import { resultCode } from "../../utils/resultCode.enum";
+import { isOwner } from "../../utils/isOwner";
 
 const schemaDeleteComment = Joi.object({
   id: Joi.string().hex().min(24).max(24).required(),
@@ -17,6 +18,15 @@ export const deleteComment = async (req: Request, res: Response) => {
   }
 
   try {
+    const commentData = await Comment.findById(id);
+
+    //valida que exista el comentario
+    if (!commentData) { return res.status(200).json({ header: { resultCode: resultCode.COMMENT_NOT_FOUND, error: 'comment not found' } }); }
+
+    //Valida que el userId sea el mismo que el id del autor
+    const userId = commentData.author.userId
+    if (await isOwner(req, res, userId)) { return res.status(401).json({ header: { resultCode: resultCode.UNAUTHORIZED, message: 'Unauthorized' } }); }
+
     await Comment.deleteOne({ _id: id });
 
     res.status(200).json({
