+++ src/controllers/post/deletePost.controller.ts
@@ -1,5 +1,6 @@
-import { Request, Response } from "express";
 import Joi from "joi";
+import { Request, Response } from "express";
+import { isOwner } from "../../utils/isOwner";
 import { Post } from "../../schemas/post.schema";
 import { resultCode } from "../../utils/resultCode.enum";
 
@@ -17,7 +18,16 @@ export const deletePost = async (req: Request, res: Response) => {
   }
 
   try {
-    const post = await Post.deleteOne({ _id: id });
+    const postData = await Post.findById(id);
+
+    //valida que el post exista
+    if (!postData) { return res.status(200).json({ header: { resultCode: resultCode.POST_NOT_FOUND, error: 'Post not found' } }); }
+
+    //Valida que el userId sea el mismo que el id del autor
+    const userId = postData.author.userId
+    if (await isOwner(req, res, userId)) { return res.status(401).json({ header: { resultCode: resultCode.UNAUTHORIZED, message: 'Unauthorized' } }) }
+
+    await Post.deleteOne({ _id: id });
 
     res.status(200).json({
       header: {
